feat: 画面の向きに応じたオーバーレイを追加し、レスポンシブデザインを改善
